nopbaitap

+++ Session16.html/Lesson6.js
@@ -1,4 +1,4 @@
-let dola= parseInt (prompt(`Nhập số đông $ cần đổi:`));
+let dola= parseFloat(prompt(`Nhập số đông $ cần đổi:`));
 let vnd=25.000;
 let result =dola * vnd;
 alert(`Số tiền đổi ra được là:` + result );
